+++ src/garminhd/GarminHDReceive.cpp
@@ -285,7 +285,7 @@ void *GarminHDReceive::Entry(void) {
       }
     }
 
-    struct timeval tv = {(long)0, (long)(MILLIS_PER_SELECT * 1000)};
+    struct timeval tv = {(int)0, (int)(MILLIS_PER_SELECT * 1000)};
 
     fd_set fdin;
     FD_ZERO(&fdin);
