+++ src/raymarine/RME120Receive.cpp
@@ -180,7 +180,7 @@ void *RME120Receive::Entry(void) {
         no_spoke_timeout = 0;
       }
     }
-    struct timeval tv = {(long)0, (long)(MILLIS_PER_SELECT * 1000)};
+    struct timeval tv = {0, (int)(MILLIS_PER_SELECT * 1000)};
 
     fd_set fdin;
     FD_ZERO(&fdin);
